refactor: emojiId형식 Long으로 통일

+++ app/src/main/java/com/issueissyu/fe/domain/model/pin/PinModels.kt
@@ -50,7 +50,7 @@ data class PinEmoji(
 )
 
 data class PinEmojiCandidate(
-    val emojiId: Int,
+    val emojiId: Long,
     val emojiImageUrl: String,
     val isDefault: Boolean,
     val isOwned: Boolean,
