+++ app/src/main/java/com/issueissyu/fe/domain/repository/PinRepository.kt
@@ -31,10 +31,10 @@ interface PinRepository {
     suspend fun getEmojiCandidates(): Result<List<PinEmojiCandidate>>
 
     /** 바텀시트(피커): POST. [emojiId]가 null이면 반응 삭제. */
-    suspend fun applyPinEmojiFromPicker(pinId: Long, emojiId: Int?): Result<Long?>
+    suspend fun applyPinEmojiFromPicker(pinId: Long, emojiId: Long?): Result<Long?>
 
     /** 목록 칩: PUT. 동일 이모지 재요청 시 백엔드에서 토글(취소) 처리. */
-    suspend fun togglePinEmojiFromList(pinId: Long, emojiId: Int): Result<Long?>
+    suspend fun togglePinEmojiFromList(pinId: Long, emojiId: Long): Result<Long?>
 
     suspend fun likePin(pinId: Long): Result<PinLike>
